docs(story): AssetDetailsPanel stories in new format
