+++ src/components/AssetDetailsPanel/stories/basic.story.mdx
@@ -1,0 +1,42 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { AssetDetailsPanel } from '../AssetDetailsPanel';
+import { AssetDetailsPanelPure } from '../AssetDetailsPanelPure';
+import { decorators } from './helper';
+
+<Meta title="Assets|AssetDetailsPanel" decorators={decorators} />
+
+# Asset Details Panel
+
+This component loads meta information for an asset and presents it as a key/value list in a table with two columns.
+The component is used as one of the panes in AssetMeta component.
+
+**NOTE:** The component should have `ClientSDKProvider` as a parent component in react component tree.
+
+## Basic usage:
+
+Imports:
+
+```jsx
+import 'antd/dist/antd.css';
+
+import React from 'react';
+import { AssetDetailsPanel } from '@cognite/gearbox';
+```
+
+<Preview>
+  <Story name="Basic Usage">
+    <AssetDetailsPanel assetId={4518112062673878} />
+  </Story>
+</Preview>
+
+## Available props
+
+<Props of={AssetDetailsPanelPure} />
+
+### Types
+
+`Asset` type can be imported from `@cognite/sdk`:
+
+```js
+import { Asset } from '@cognite/sdk';
+```
